+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Improves Error Deserialization to dynamically handle edge cases that have a bad format
+
 ## v6.4.0 (2023-04-04)
 
 - Add `getNextPage` function that retrieves the next page of results for a paginated collection